Display description for pending specs
There are various reasons why a spec might be pending. With the
support for skipping specs from the command line, there are now at
least two: Marked as pending, or skipped. This change enables
buttercup to display the reason for why a test was not run.

Users can now use this functionality to signal buttercup-pending
with an argument to show why something was skipped.

Fixes #22