Add assume form to cancel tests if conditions fail
The assume form cancels the current test if the given condition is not
met.  It is intended to be used in tests that depend on a certain
global state that might not be present, e.g. a tool in "exec-path" or a
network connection, etc.

Example:

(describe "Eggs"
  (it "gets eggs from the internet"
    (assume (network-available-p) "No network connection")
    (my/get-eggs)))

Output if the assumption is not met

Eggs
  gets eggs from the internet  !! CANCELLED !! No network connection