Fix several edge cases in "spy-on"
The "spy-on" function now works on commands and autoloaded functions,
and checks for the right number of arguments for each modifier.

For commands, it preserves the interactive form of the original
command, unless the given replacement overrides it (in which case a
warning is issued). This means that the spy can be executed as a
command in the same way as the original.

For autoloads, "spy-on" simply forces the load to happen so that the
function becomes an ordinary function instead of an autoloaded one.

Finally, all modifiers now check whether the correct number of args
was provided instead of silently ignoring extra ones. In addition, the
argument for ":and-throw-error" now defaults to "error" if a more
specific error type is not specified.

Tests are added for all three features.

This should fix #56 and #103.

Add autoload test