Allow ERT "should" and similar to work with buttercup
This defines a new macro "buttercup-with-converted-ert-signals" that
converts ERT failure and test-skipping signals to the equivalent
buttercup ones. This allows "should" and other ERT test functions to
work within buttercup test specs, thus making it easier for people to
port ERT tests to buttercup.

The test for the macro expansion of "it" has been modified for the new
expansion, and a few tests have been added to verify the new
functionality.