Implement closure-based expect macro
This redefines the "expect" macro to wrap all its arguments (except
the matcher) in closures. Matchers are now expected to accept all
arguments as closures and must "funcall" them to obtain their value.
This allows matchers to also obtain the unevaluated expression for
each of their arguments in order to generate more informative error
messages. All built-in matchers are reimplemented to conform to this
requirement. This fixes #54.

In addition, this adds some convenience macros for defining matchers.
The macro "buttercup-define-matcher-for-unary-function" automatically
generates a matcher for any predicate function, while
"buttercup-define-matcher-for-binary-function" does the same for a
comparison function, and also automatically uses any explainer defined
by ERT for that function for more precise messages in case of a
mismatch. This fixes #28. More generally, the
"buttercup--test-expectation" macro makes it easier to construct the
non-intuitive return values required by buttercup matchers. This fixes

the change to closure-based matchers is mostly backward-compatible,
with the only user-visible diferences being better error messages.
However, one major backward-incompatible change is the handling of the
":to-throw" matcher, which no longer requires its argument to be
wrapped in a lambda form, since the "expect" macro now does so
automatically for all matchers. This fixes #41. In order to aid
migration, it would be helpful to add a warning when a function
returns another function, which is a likely indicator that the desired
expression was wrongly wrapped in a lambda form and wasn't evaluated.

Another new feature is support for issuing warnings during tests. Any
warning issued during a test is collected in a special buffer and
displayed after the test finishes.

Tests are added for all new functionality, and existing tests are
adapted to the new code.