Fix the buttercup-run tests
Let-binding the buttercup-reporter in the suite did not work, the
suites and specs used as test data were reported by the main
buttercup-reporter.  Move the let-binding into the specs instead.

Use cl-flet to define local reporter and runner functions.

Make the runner function local to the final spec for buttercup-run.