Really fix the buttercup-run tests
The fix in 672bce1276c47e7c09eae66593168b83f445a8c7 was not enough.
`buttercup-suites' was not correctly scoped, and somehow the global
value was overwritten.  The result being an incorrect number of run
tests being reported at the end of the test run.
I have no clue why let-binding `buttercup-suites' in each spec works
better, but it does.

To get it all to work on Emacs 24.3 I had to drop some of the cl-flet
for less elegant solutions.