Properly handle return values and thrown signals in spies
Attempting to access the return valye of a spy context representing a
call that threw a signal now throws an error. Similarly, attempting to
access the thrown signal of a successful function call also throws an
error.

This is implemented by using two new structs derived from the
spy-context struct: one for successful calls and one for errors.
Ideally, we would use cl-defgeneric, but it is not available prior to
Emacs 25.1.

Tests are added for the new behavior, and several existing tests are
modified to accommodate it.