Raise an error if spy-on is used in the wrong context
`spy-on' is only allowed inside a `buttercup-with-cleanup'
environment.  Ensure this by raising errors when used wrong.  Some of
these errors will be signaled while reading the tests, but some (like in
`before-each') will be signaled when the tests are run.

The `buttercup--cleanup-functions' variable gets a new default value
that does not satisfy `listp'.  This non-list value is used to detect
when `spy-on' is called in an invalid context.

Fixes #122.