Optionally do not print skipped tests
When `buttercup-silent-skipping' is non-nil, do not print specs that
were marked as skipped before `buttercup-run' was started.  Also do
not print the descriptions for suites that only contain such skipped
specs.

Specs that are marked as skipped by `assume' will still be printed.

Add an option '--silent-skipping' to bin/buttercup.