test: Verify output in the "The batch reporter" suite
Use buttercup--print-collector to collect the output as it would
appear on the terminal and verify that it matches expectation.

The tests used to verify the calls to buttercup--print, but this is
fragile at best.

"should print the elasped time for all specs" is removed as the
elapsed time is checked in all other tests.