test: Verify full line for reporter spec-done tests
The batch reporter splits output of specs between spec-started and
spec-done, and in the spec-done tests the full line has to be verified
to make sure that any control characters (\r,\t,\b,\n ...) has been
used correctly.  spec-started still has to be verified separately for
text that should be displayed before the test is actually run, as the
spec description shall be displayed while running the test.