Fix stack frame collection in buttercup--backtrace
Changes to the inner workings of buttercup has made the condition for
when to stop collecting frames out-of-date.  Looking for
buttercup--funcall is no longer enough; buttercup--expr-and-value and
buttercup--apply-matcher must alse be used.  The number of frames to
discard depends on which of these functions is found.

The frame inspection loop is rewritten using cl-do*, which allows for
early exit with cl-return.