Store data about failed loads as failed specs
For each failed file load in buttercup-dun-discover, store the failure
and stack in a pre-failed spec.  Add the specs to a suite and add it
to the end of buttercup-suites.

This means that failed file load are displayed clearly, with error
message and backtrace, and buttercup-run-discover will signal an error
as for any other failing test.

Fixes #!07.