Prevent early backtrace on test failure
The buutercup finctions called from the commandline uses error to make
Emacs return a non-zero exit code.  This will trigger early backtrace,
which is not useful to buttercup.  Disable it by setting
backtrace-on-error-noninteractive at the start of these functions.

It could be that backtrace-on-error-noninteractive should be set just
before the error is signaled, but then it is hard to decide whether
this is a commandline batch run.

Fixes #214.