Check position of the :var and :var* keywords
Signal an error if the :var or :var* keyword is found in any other
position than the first argument of a describe macro.  Fixes #223.

There are several issues with the :var(*) keywords, the most
important being it's inability to handle dynamic variables (#127).
But there is also the question of what the correct behaviour would be
if you use multiple :var(*) or when they are put anywhere but at the
beginning of the describe macro.