Extract buttercup--handle-to-throw from :to-throw matcher
It is very hard to test the :to-throw matcher due to the problems with
nested debuggers.  Split the code that doesn't actually have to deal
with the raising and catching of signals into a separate function that
can actually be tested.

The new function have a new `let' encompassing the already existing
`let*'.  This is to minimize the commit diff.