Add function buttercup--simple-format
buttercup--simple-format is a simplified variant of format-spec, but
it supports functions as substitutions which was only added in Emacs
29.

Compared to format-spec it drops the IGNORE-MISSING and SPLIT
parameters and just ignores any unknown specifications, including % at
the end of line or not followed by a letter.

The specifications can not have flags, width or precision.  This is a
simple search and replace.

As a consequence of these simplifications the code is much smaller and
should actually run faster than format-spec.

It cannot completely replace buttercup-format-spec, as users that
define their own matcher may use the flags, width and precision.