Also bind and declare comp-enable-subr-trampolines
The variable native-comp-enable-subr-trampolines was called
comp-enable-subr-trampolines in Emacs 28.  It's fine to set both to
nil as long as they are declared as dynamic variables.

But comp-enable-subr-trampolines is an obsolete alias in Emacs 29, use
with-suppressed-warnings to silence it.

But with-suppressed-warnings wasn't added until Emacs 27.  Make
buttercup-with-cleanup more complicated by only using
with-suppressed-warnings when it is defined.
comp-enable-subr-trampolines is just a name if
with-suppressed-warnings is not available, so using it in the
let-statement is fine.

For some reason it was not possible to use
(boundp 'comp-enable-subr-trampolines)
to select the variable to bind, even with (require 'comp).

Fixes #236.