Drop support for Emacs 24.3
Emacs 24.3 was released in March of 2013.  Emacs 24.4 was released in
October of 2014.  emacs-buttercup tries to support as old Emacsen as
possible as it is a tool that can reasonably be expected to be used on
old Emacsen during testing of other packages.  But there must be a
limit.  Emacs 24.3 is 10.5 years old at this point, and all remotely
recent distributions come with newer versions of Emacs.