test: Update tests to work with OClosures
When `buttercup--wrap-expr' uses `buttercup-thunk' OClosures, it
actually returns a form that has to be `eval'ed to get a
`buttercup-thunk'.
This is not an issue when `buttercup--wrap-expr' is used in the
`expect' macro, because the expansion of `expect' will be read/eval:ed
anyway.
But in the tests the return will sometimes have to be explicitly
evaled before it's processed by other functions.