Avoid printing double lines in GitHub Actions
The "terminal" used in github action does not support the use of
carriage return '\r' to move back to column zero and overwrite the
line.  Carriage return will instead move to the next line.

The consequence for the buttercup batch reporer was that each spec
description was printed twice, once without color before the spec was
run and once with color once the spec had run and the result was
known.  Reuse the same mechanism as for specs with descriptions
containing newlines, don't print the spec description before running
it unless color is disabled.

Fixes #181.