tests: Add backtrace tests for function matchers
Function matchers (expect A (lambda (x y z) (< x y z)) B C) is an
obscure feature of buttercup. An interesting fact is that the wrapped
arguments are evaluated _before_ the matcher is called in contrast to
matchers defined with buttercup-define-matcher where the matcher is
assumed to do that itself.