test: Set b*c*-r*r-batch--start-time in with-local-buttercup
The value of the actual buttercup-reporter-batch--start-time variable
should only be used by the main test loop, not by the tests used to
verify buttercup.  Otherwise the time could easily be accidentaly
reset by the individual tests.  The variable value needs to be an
actual timestamp or some elapsed times will be reported as "-0.00ms"
which will cause test failures.

The "on the buttercup-started event" and "on the buttercup-done event"
suites should not use the :var tag to set the variable.  This is not
supported for dynamic variables, which was exposed by the recently
added warning.